revert: revert last two commits related to calendar day selection changes
